pyproject.toml
[project]
name = "test-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.0",
    "fastapi>=0.116.2",
    "pydantic-settings>=2.10.1",
    "pyodbc>=5.2.0",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
